refactor(orchestration): extract NoticeDeliveryService
First of the nine leaf services. Sixteen methods, 354 body lines, moved
verbatim. 3426 -> 3115 lines. Fourteen become one-line facade delegations;
canInjectGroupIntoCoordinator and canInjectTaskIntoCoordinator have no caller
outside the sixteen and stay private on the service.

The plan's port list was wrong again. It named `logger`, but `this.deps.logger`
appears zero times anywhere in orchestration-service.ts -- every log goes
through kernel.logEvent, and the kernel holds the logger. The Pick is three
ports wide: now, loadState, saveState. The isolation test constructs the
service from a bare object literal of exactly those three, which is the point:
a service that cannot be built without ports it never calls has not been split.

The plan's smoke test was also wrong in three ways -- it passed an input
object to markTaskNoticeDelivered (two positional args), never awaited
listPendingTaskNotices, and put `logger` in the deps literal. It was rewritten
from the real signatures.

Seven of the sixteen open kernel.mutate. None was added, removed, or relocated.
